Refactored Code into multiple files
