+++ C-Sharp-Task-2/AccountType.cs
@@ -1,0 +1,9 @@
+namespace FinancialSystem
+{
+    public enum AccountType
+    {
+        Savings,
+        Checking,
+        Investment
+    }
+}
