+++ C-Sharp-Task-2/CheckingAccount.cs
@@ -1,0 +1,21 @@
+namespace FinancialSystem
+{
+    public class CheckingAccount : Account
+    {
+        private const double MonthlyFee = 10.0;
+
+        public CheckingAccount() : base(0, AccountType.Checking)
+        {
+        }
+
+        public override void ApplyInterestOrFees()
+        {
+            Balance -= MonthlyFee;
+        }
+
+        public override void PrintBalance()
+        {
+            Console.WriteLine($"Your CheckingAccount balance is {Balance:C}\n");
+        }
+    }
+}
