+++ C-Sharp-Task-2/Program.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Text.Json;
+
+namespace FinancialSystem
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            var app = new FinancialSystem();
+            app.Start();
+        }
+    }
+}
